# Add an ASCII fast-pass to ROW